Prepare support for more than one ffi descriptor
This is probably a bit of a dead-end. Ultimately the Rust side will need
to manage connections which means the C side will need to hold on to
some sort of Manager struct that it uses to create new descriptors. At
that point, why pass the descriptor back at all when you could pass an
identifier to the manager? The FFI layer wil likely remain but having
the C side hold on to the descriptors will probably go away