Slack integration
Introduce a manager for slack connections. It holds on to descriptor ids
by channel id for input and just writes directly to the channel for output.

Manager hackily monitors for when the C-side has requested a new
descriptor. This would be better if the C-side provided a list of the
descriptor ids it already has but I don't want to figure that out on
that side.