Introduce mud-comms crate
Empty lib crate that produces a staticlib.

> This format is recommended for use in situations such as linking Rust code into an existing non-Rust application because it will not have dynamic dependencies on other Rust code.

https://doc.rust-lang.org/reference/linkage.html

---

Add cargo/rustc checking to `configure`
(https://ljones.dev/blog/gsoc-2/)

`$ autoconf -I cnf -o configure cnf/configure.in` to generate a new
`configure` script.

---

Modify `Makefile` to build and link the Rust staticlib when it is
changed (add it as a dependency for the `circle` binary and add the
manifest and source files for `mud-comms` as a dependency for the static
lib)

`./configure` to generate the Makefile. `make -C src` to build.