Replace raw socket descriptor with descriptor abstraction
This is a big diff, and almost unreadable.

The main objective is to remove all Telnet code and hide it behind the
FFI layer so we can swap it out later.

That means:
* Replacing the "mother descriptor" with a general "descriptor manager"
  * the manager now handles listener init and detecting new connections
  * new_descriptor will return a pointer if there is a new connection
* better names (ie `new_descriptor_data` since it creates the
bookkeeping struct that CircleMUD uses, not the actual socket
descriptor, similarly `close_socket` which was also confusing with
`CLOSE_SOCKET`)
* Lots of Telnet/socket helper functions are gone