Replace DescriptorManager/Descriptor /w trait
There is a pretty clear relationship between the DescriptorManager and
Descriptor - the manager owns creating/waiting on new connections and
the descriptor provides the reading/writing handle plus additional
bookkeeping.

That behaviour can be represented with a trait making it easier to
introduce a new connection manager.

(This also introduces a `get_hostname` method to abstract the direct
access to the Descriptor previously)