Implement `socket_std::SocketDescriptorManager::block_until_descriptor`
`block_until_descriptor` would return immediately causing the game loop
to busy wait. Fixing this is not straightforward since `TcpListener`
does not provide methods to check for new streams (`accept` and
`incoming` immediately provide the new stream) but only want to know
*if* there is a new stream. We also cannot change the game loop to block
on `new_descriptor` - then the loop would be blocked for existing
players while waiting for a new connection.

This approache moves the `TcpListener` into a thread that sends new
streams into a channel and updates a conditional variable every time it
does so. `new_descriptor` now receives streams off the channel and also
updates the conditional variable when it does so.
`block_until_descriptor` can now lock and wait for for the conditional
variable to change`block_until_descriptor` would return immediately
causing the game loop to busy wait. Fixing this is not straightforward
since `TcpListener` does not provide methods to check for new streams
(`accept` and `incoming` immediately provide the new stream) but only
want to know *if* there is a new stream. We also cannot change the game
loop to block on `new_descriptor` - then the loop would be blocked for
existing players while waiting for a new connection.

This approach moves the `TcpListener` into a thread that sends new
streams into a channel and updates a conditional variable every time it
does so. `new_descriptor` now receives streams off the channel and also
updates the conditional variable when it does so.
`block_until_descriptor` can now lock and wait for for the conditional
variable to change. While convoluted this results in the least change to
the game loop (aka none).