Introduce Slack manager and descriptor
Slack integration involves exposing an events endpoint with variable
behaviour (initially it is verified before events are pushed) in
addition to the normal OAuth and API authorization to send messages to
users.

WebSockets are also possible with Slack but early versions of
`slack-morphism` only supported push events.

TODO: module is a mess